requirements.txt
flask
flask_restful
flask-mongoengine
mongoengine_goodjson
flask-cors
